requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
Django==5.1.3
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
idna==3.10
Markdown==3.7
openpyxl==3.1.5
pillow==11.1.0
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.4.0
xmltodict==0.14.2
